refactor: split dashboard view builders
